+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+* Added `material.OperationXYZ` for merged XYZ operations.
+
+### Changed
+* Renamed `shader_database.Dependency` to `shader_database.Value` and updated associated type names.
+* Renamed `shader_database.BufferDependency` to `shader_database.Parameter` and updated getter.
+* Renamed `material.OutputExprXyz.assignment` to `material.OutputExprXyz.expr`.
+* Renamed `material.OutputExprXyz.assignments` to `material.OutputExprXyz.exprs`.
+
+### Removed
+* Removed `material.Assignment` and associated types and replaced with `shader_database.OutputExpr`.
+
+
 ## 0.16.0 - 2026-05-26
 ### Added
 * Added support for wimdo export for Xenoblade Chronicles X Definitive Edition.